feat(grpc) add grpc/grpcs nginx locations
Extend the Kong 'Access' phase to detect requests to gRPC-based
services and to internally redirect to `grpc_pass` instead of
`proxy_pass`.

Separate @grpc and @grpcs locations are defined for plaintext and
secure gRPC services.

Additionally, this commit ensures the request context is
restored/restashed and avoids re-execution of certain phases:
- Restore stashed context upon entry in gRPC (done in the rewrite phase,
after the internal redirection)
- Avoid re-running already executed handlers; this is done in different
ways in the rewrite and pre-access handlers:
    * rewrite: inspect `ngx.var.kong_proxy_mode`; if it is gRPC, return
    early
    * pre-access (access.before): inspect `ngx.ctx.service`; if it is
    non-nil and a gRPC service, return early

Co-authored-by: Luke Gorrie <luke@snabb.co>